Deployment boy v1.0.0

+++ Ganga/client/src/Components/Panels/AdminPanel/productList.jsx
@@ -1,16 +1,20 @@
 import React, { useState } from "react";
 import { Link } from "react-router-dom";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { DataGrid } from "@mui/x-data-grid";
 import { TiDeleteOutline } from "react-icons/ti";
+import { BsPencilSquare } from "react-icons/bs";
+import Swal from 'sweetalert2';
 
 import { deleteProduct } from "../../Redux/Actions/actions";
 import ProductsChart from "./aCharts/Products";
 import s from "./admin.module.css";
 
 export default function ProductList({ products }) {
-
   const dispatch = useDispatch();
+  const categories = useSelector((state) => state.categories);
+  const subcategories = useSelector((state) => state.dbSubcategories);
+  products = products.filter(p => p.approved === true);
 
   const columns = [
     { field: "id", headerName: "ID", width: 50 },
@@ -29,7 +33,7 @@ export default function ProductList({ products }) {
         return (
           <>
             <Link to={"/product/" + id}>
-              <button className={s.editar}> edit </button>
+              <button className={s.editar}> <BsPencilSquare /> </button>
             </Link>
             <button onClick={() => handleDelete(id)}>
               {" "}
@@ -42,18 +46,53 @@ export default function ProductList({ products }) {
   ];
 
   function handleDelete(id) {
-    dispatch(deleteProduct(id));
-    setRows(rows.filter((i) => i.id !== id));
-    window.location.reload();
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: "Se borraran todos los datos del producto.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Confirmar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Eliminado!',
+          'El producto ha sido eliminado.',
+          'success'
+        )
+        dispatch(deleteProduct(id));
+        setRows(rows.filter((i) => i.id !== id));
+        window.location.reload();
+      }
+    })
   }
 
   let Rows = products?.map((p) => {
+    if (p.categories === null) {
+      p.categories = categories.filter(c => c.id === p.categoryId);
+      p.categories = p.categories[0].name;
+    }
+
+    if (p.subcategoryId > 0) {
+      let sub = subcategories.filter(s => s.id === p.subcategoryId);
+      sub = sub[0]?.name
+      p.subcategories = sub;
+    }
+
+    if (p.brand === null || p.brand === undefined) {
+      p.brand = "indefinida"
+    }
+
     return {
       id: p.id,
       Nombre: p.name,
       Precio: "$" + p.price,
       Categoria: p.categories,
+      Subcategoria: p.subcategories,
       Stock: p.stock,
+      Marca: p.brand
     };
   });
