+++ Ganga/client/src/Components/Panels/AdminPanel/userList.jsx
@@ -3,6 +3,8 @@ import { useDispatch } from "react-redux";
 import { Link } from "react-router-dom";
 import { DataGrid } from "@mui/x-data-grid";
 import { TiDeleteOutline } from "react-icons/ti";
+import { BsPencilSquare } from "react-icons/bs";
+import Swal from 'sweetalert2';
 
 import NewUsersChart from "./aCharts/NewUsers";
 import s from "./admin.module.css";
@@ -32,7 +34,7 @@ export default function UserList({ users }) {
         return (
           <>
             <Link to={"/user/" + id}>
-              <button className={s.editar}> edit </button>
+              <button className={s.editar}><BsPencilSquare /></button>
             </Link>
             <button onClick={() => handleDelete(id)}>
               {" "}
@@ -58,9 +60,27 @@ export default function UserList({ users }) {
   const [rows, setRows] = useState(Rows);
 
   function handleDelete(id) {
-    dispatch(deleteUser(id));
-    setRows(rows.filter((i) => i.id !== id));
-    window.location.reload();
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: "Se borraran todos los datos del usuario.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Confirmar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Eliminado!',
+          'El usuario ha sido eliminado.',
+          'success'
+        )
+        dispatch(deleteUser(id));
+        setRows(rows.filter((i) => i.id !== id));
+        window.location.reload();
+      }
+    })
   }
 
   return (
