+++ Ganga/client/src/Components/Panels/AdminPanel/vendorList.jsx
@@ -3,6 +3,8 @@ import { useDispatch } from "react-redux";
 import { DataGrid } from "@mui/x-data-grid";
 import { Link } from "react-router-dom";
 import { TiDeleteOutline } from "react-icons/ti";
+import { BsPencilSquare } from "react-icons/bs";
+import Swal from 'sweetalert2';
 
 import RevenueChart from "./aCharts/Revenue";
 import s from "./admin.module.css";
@@ -28,7 +30,7 @@ export default function VendorList({ vendors, products }) {
         return (
           <>
             <Link to={"/user/" + id}>
-              <button className={s.editar}> edit </button>
+              <button className={s.editar}> <BsPencilSquare /> </button>
             </Link>
             <button onClick={() => handleDelete(id)}>
               {" "}
@@ -41,8 +43,27 @@ export default function VendorList({ vendors, products }) {
   ];
 
   function handleDelete(id) {
-    dispatch(deleteUser(id));
-    setRows(rows.filter((i) => i.id !== id));
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: "Se borraran todos los datos del vendedor.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Confirmar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Eliminado!',
+          'El vendedor ha sido eliminado.',
+          'success'
+        )
+        dispatch(deleteUser(id));
+        setRows(rows.filter((i) => i.id !== id));
+        window.location.reload();
+      }
+    })
   }
 
   const Rows = vendors?.map((u) => {
