+++ Ganga/client/src/Components/Panels/AdminPanel/vendorProducts.jsx
@@ -4,6 +4,8 @@ import { DataGrid } from "@mui/x-data-grid";
 import { useDispatch, useSelector } from "react-redux";
 import { getProduct } from "../../Redux/Actions/actions";
 import { TiDeleteOutline } from "react-icons/ti";
+import Swal from 'sweetalert2';
+import { BsPencilSquare } from "react-icons/bs";
 
 import { deleteProduct } from "../../Redux/Actions/actions";
 
@@ -14,6 +16,8 @@ export default function VendorProducts({ id }) {
 
     const products = useSelector((state) => state.product);
     const vendorProducts = products.filter((p) => p.owner === id);
+    const categories = useSelector((state) => state.categories)
+    const subcategories = useSelector((state) => state.dbSubcategories)
 
     useEffect(() => {
         dispatch(getProduct());
@@ -38,7 +42,7 @@ export default function VendorProducts({ id }) {
                 return (
                     <>
                         <Link to={"/product/" + id}>
-                            <button className={s.editar}> edit </button>
+                            <button className={s.editar}> <BsPencilSquare /> </button>
                         </Link>
                         <button onClick={() => handleDelete(id)}>
                             {" "}
@@ -51,18 +55,52 @@ export default function VendorProducts({ id }) {
     ];
 
     function handleDelete(id) {
-        dispatch(deleteProduct(id));
-        setRows(rows.filter((i) => i.id !== id));
-        window.location.reload();
+        Swal.fire({
+            title: 'Estas seguro?',
+            text: "Se borraran todos los datos del producto.",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            cancelButtonText: 'Cancelar',
+            confirmButtonText: 'Confirmar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire(
+                    'Eliminado!',
+                    'El producto ha sido eliminado.',
+                    'success'
+                )
+                dispatch(deleteProduct(id));
+                setRows(rows.filter((i) => i.id !== id));
+                window.location.reload();
+            }
+        })
     }
 
     let Rows = vendorProducts?.map((p) => {
+        const MyCategory = categories.filter(c => c.id === p.categoryId);
+        const myCategory = MyCategory[0].name;
+        const mySubcategory = subcategories.filter(s => s.id === p.subcategoryId);
+        console.log("mySubcategory", mySubcategory);
+
+        let mySubCategory = "";
+        if (mySubcategory?.length === 0) {
+            mySubCategory = "no definida"
+
+        } else {
+            mySubCategory = mySubcategory[0].name;
+
+        }
+
         return {
             id: p.id,
             Nombre: p.name,
             Precio: "$ " + p.price,
-            Categoria: p.categories,
+            Categoria: myCategory,
+            Subcategoria: mySubCategory,
             Stock: p.stock,
+            Marca: p.brand
         };
     });
