+++ Ganga/client/src/Components/Panels/VendorPanel/vProductList.jsx
@@ -3,7 +3,10 @@ import { Link } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import { DataGrid } from "@mui/x-data-grid";
 import { TiDeleteOutline } from "react-icons/ti";
-import { deleteProduct, getCategories, getDbSubcategories } from "../../Redux/Actions/actions";
+import { BsPencilSquare } from "react-icons/bs";
+import Swal from 'sweetalert2';
+
+import { deleteProduct, getCategories, getDbSubcategories, allReviews } from "../../Redux/Actions/actions";
 
 import s from "../AdminPanel/admin.module.css";
 
@@ -13,8 +16,9 @@ export default function VendorProductList({ products, user }) {
     const subcategories = useSelector((state) => state.dbSubcategories)
     let myProducts = products.filter((p) => p.owner === user.id);
 
-    console.log("pp", products);
-    console.log("sub", subcategories);
+    useEffect(() => {
+        dispatch(allReviews())
+    }, [dispatch]);
 
     useEffect(() => {
         dispatch(getCategories());
@@ -23,6 +27,8 @@ export default function VendorProductList({ products, user }) {
     useEffect(() => {
         dispatch(getDbSubcategories());
     }, [dispatch]);
+
+
 
     const columns = [
         { field: "id", headerName: "ID", width: 50 },
@@ -41,7 +47,7 @@ export default function VendorProductList({ products, user }) {
                 return (
                     <>
                         <Link to={"/product/" + id}>
-                            <button className={s.editar}> edit </button>
+                            <button className={s.editar}> <BsPencilSquare /> </button>
                         </Link>
                         <button onClick={() => handleDelete(id)}>
                             {" "}
@@ -54,26 +60,44 @@ export default function VendorProductList({ products, user }) {
     ];
 
     function handleDelete(id) {
-        dispatch(deleteProduct(id));
-        setRows(rows.filter((i) => i.id !== id));
-        window.location.reload();
+        Swal.fire({
+            title: 'Estas seguro?',
+            text: "Se borraran todos los datos del producto.",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            cancelButtonText: 'Cancelar',
+            confirmButtonText: 'Confirmar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                Swal.fire(
+                    'Eliminado!',
+                    'Tu producto ha sido eliminado.',
+                    'success'
+                )
+                dispatch(deleteProduct(id));
+                setRows(rows.filter((i) => i.id !== id));
+                window.location.reload();
+            }
+        })
     }
 
     let Rows = myProducts?.map((p) => {
         const MyCategory = categories.filter(c => c.id === p.categoryId);
         const myCategory = MyCategory[0].name;
         const mySubcategory = subcategories.filter(s => s.id === p.subcategoryId);
-        console.log("subc", mySubcategory);
+
         let mySubCategory = "";
         if (mySubcategory?.length === 0) {
             mySubCategory = "no definida"
-            console.log("msubc1", mySubCategory);
+
         } else {
             mySubCategory = mySubcategory[0].name;
-            console.log("msubc2", mySubCategory);
+
         }
 
-        console.log("msubc3", mySubCategory);
+
         return {
             id: p.id,
             Nombre: p.name,
